refactor : #409 Participation 도메인 분리
