+++ src/main/resources/data.sql
@@ -66,9 +66,13 @@ VALUES (1, '2023-3-10-11-0', '2023-3-10-11-0', '테스트 대댓글1', 1, 1);
 INSERT INTO reply_comment (reply_comment_id, created_date_time, last_modified_date_time, body, comment_id, member_id)
 VALUES (2, '2023-3-10-11-0', '2023-4-10-11-0', '테스트 대댓글2', 1, 1);
 
--- 참여 2번 게시글에 1번 유저가 참여한 상태
+-- 참여상태
 INSERT INTO participation(member_id, recruitment_board_id)
 VALUES (1, 1);
+INSERT INTO participation(member_id, recruitment_board_id)
+VALUES (2, 1);
 
+INSERT INTO participation(member_id, recruitment_board_id)
+VALUES (3, 1);
 INSERT INTO participation(member_id, recruitment_board_id)
 VALUES (2, 2);
