+++ src/test/java/mokindang/jubging/project_backend/recruitment_board/domain/participation/ParticipationTest.java
@@ -1,6 +1,7 @@
 package mokindang.jubging.project_backend.recruitment_board.domain.participation;
 
 import mokindang.jubging.project_backend.member.domain.Member;
+import mokindang.jubging.project_backend.participation.domain.Participation;
 import mokindang.jubging.project_backend.recruitment_board.domain.RecruitmentBoard;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.extension.ExtendWith;
